+++ src/game/nlcore/action/actionTypes.ts
@@ -234,6 +234,7 @@ export type LayerActionContentType = {
 /* Video */
 export const VideoActionTypes = {
     action: "video:action",
+    preload: "video:preload",
     show: "video:show",
     hide: "video:hide",
     play: "video:play",
@@ -245,7 +246,7 @@ export const VideoActionTypes = {
 export type VideoActionContentType = {
     [K in typeof VideoActionTypes[keyof typeof VideoActionTypes]]:
     K extends "video:action" ? any :
-    K extends "video:show" | "video:hide" | "video:play" | "video:pause" | "video:stop" | "video:resume" ? [] :
+    K extends "video:preload" | "video:show" | "video:hide" | "video:play" | "video:pause" | "video:stop" | "video:resume" ? [] :
     K extends "video:seek" ? [number] :
     any;
 }